+++ internal/cluster/clusterpb/MetaStateSnapshot.go
@@ -401,8 +401,28 @@ func (rcv *MetaStateSnapshot) IndexGroupsLength() int {
 	return 0
 }
 
+func (rcv *MetaStateSnapshot) PlacementGenerations(obj *PlacementGenerationEntry, j int) bool {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(42))
+	if o != 0 {
+		x := rcv._tab.Vector(o)
+		x += flatbuffers.UOffsetT(j) * 4
+		x = rcv._tab.Indirect(x)
+		obj.Init(rcv._tab.Bytes, x)
+		return true
+	}
+	return false
+}
+
+func (rcv *MetaStateSnapshot) PlacementGenerationsLength() int {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(42))
+	if o != 0 {
+		return rcv._tab.VectorLen(o)
+	}
+	return 0
+}
+
 func MetaStateSnapshotStart(builder *flatbuffers.Builder) {
-	builder.StartObject(19)
+	builder.StartObject(20)
 }
 func MetaStateSnapshotAddNodes(builder *flatbuffers.Builder, nodes flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(0, flatbuffers.UOffsetT(nodes), 0)
@@ -504,6 +524,12 @@ func MetaStateSnapshotAddIndexGroups(builder *flatbuffers.Builder, indexGroups f
 	builder.PrependUOffsetTSlot(18, flatbuffers.UOffsetT(indexGroups), 0)
 }
 func MetaStateSnapshotStartIndexGroupsVector(builder *flatbuffers.Builder, numElems int) flatbuffers.UOffsetT {
+	return builder.StartVector(4, numElems, 4)
+}
+func MetaStateSnapshotAddPlacementGenerations(builder *flatbuffers.Builder, placementGenerations flatbuffers.UOffsetT) {
+	builder.PrependUOffsetTSlot(19, flatbuffers.UOffsetT(placementGenerations), 0)
+}
+func MetaStateSnapshotStartPlacementGenerationsVector(builder *flatbuffers.Builder, numElems int) flatbuffers.UOffsetT {
 	return builder.StartVector(4, numElems, 4)
 }
 func MetaStateSnapshotEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT {